Refactor Delete Confirmation Modal
