+++ lib/src/features/bookings/presentation/screens/edit_booking_screen/widgets/form_fields/booking_type_selector_field.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:keeley/src/features/bookings/domain/objects/booking_type.dart';
 import 'package:keeley/src/features/bookings/presentation/controllers/edit_booking_controller.dart';
-import 'package:keeley/src/common_widgets/toggle_button_group.dart';
+import 'package:keeley/src/common/widgets/toggle_button_group.dart';
 import 'package:keeley/src/constants/keys.dart';
 import 'package:keeley/src/constants/strings.dart';
